Adding Hinput files
Adding modified code and common files to allow for an input of H at upper levels of model, to simulate H in-flux from space.

Modifications made to common file (com2d... .h), main, solv subroutine, and Run script.

Currently, H input value is hard-coded into solv, but turning H input on/off is controlled by flags set in Run script and used in main and solv.

This should be modified further so that H intput value can be specified in Run script.  That could be done in a similar way to how on/off is currently done using files containing flag values.