Correction/addition to previous update
Update immediately preceding this one changed run length to be user-defined, but did not correctly transfer that length to the model when run.

Added lines in the Run script to put the run length in days into the startYear.txt temporary file.

Added lines in control read-in subroutine to read in this length in days to run the model.